requirements.txt
asgiref==3.8.1
cramjam==2.10.0
decouple==0.0.7
Django==5.2
django-cors-headers==4.7.0
django-filter==25.1
django-otp==1.6.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
duckdb==1.2.2
et_xmlfile==2.0.0
fastparquet==2024.11.0
fsspec==2025.3.2
future==1.0.0
inflection==0.5.1
mysql==0.0.3
mysqlclient==2.2.7
numpy==2.2.4
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.1.0
psycopg2==2.9.10
pyarrow==19.0.1
PyJWT==2.9.0
pyotp==2.9.0
PyQt6==6.9.0
PyQt6-Qt6==6.9.0
PyQt6_sip==13.10.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
serializers==0.2.4
six==1.17.0
sqlparse==0.5.3
sqlshell==0.1.9
tzdata==2025.2
uritemplate==4.1.1
xlrd==2.0.1
